+++ app/Http/Controllers/TeamController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Service\Table;
 use App\Team;
 use App\User;
 use Illuminate\Http\Request;
@@ -10,10 +11,48 @@ use App\Http\Requests;
 
 class TeamController extends Controller
 {
+  public function getTable(Request $request) {
+    $query = Team::getAll();
+
+    $table = new Table($request);
+    $table->setBulkActions(true);
+    $table->setDateFormat('d/m/Y H:i');
+    $table->setPrimaryKey('id');
+    $table->addColumn('name', 'Nom');
+    $table->addColumn('count_user', "Nbr. de collaboratuers", function ($entity) {
+      return $entity->users->count();
+    });
+    $table->addColumn('description', "Description", function ($entity) {
+      return $entity->description != '' ? $entity->description : '---';
+    });
+    $table->addColumn('created_at', "Créée le");
+
+    // define table actions
+    $table->addAction('edit', [
+      'icon' => 'fa fa-pencil',
+      'label' => 'Modifier',
+      'route' => ['name' => 'team.form', 'args' => ['id' => '[id]']],
+      'attrs' => [
+        'chm-modal'=> '',
+        'chm-modal-options'=> '{"form":{"attributes":{"id":"teamForm", "target-table":"[chm-table]"}}}',
+      ],
+      'bulk_action' => false,
+    ]);
+    $table->addAction('delete', [
+      'icon' => 'fa fa-trash',
+      'label' => 'Supprimer',
+      'callback' => 'Team.delete',
+      'bulk_action' => true,
+    ]);
+
+    // render the table
+    return $table->render($query);
+  }
+
+
   public function index()
   {
-    $teams = Team::getAll()->paginate(10);
-    return view('teams.index', compact('teams'));
+    return view('teams.index');
   }
 
   public function form(Request $request)
@@ -62,11 +101,24 @@ class TeamController extends Controller
   /**
    * Remove the specified resource from storage.
    */
-  public function delete($tid)
+  public function delete(Request $request)
   {
-    $team = Team::find($tid);
-    $team->delete();
-    return ["status" => "success", "message" => "L'équipe a été supprimée avec succès !"];
+    if (empty($request->ids)) return;
+
+    foreach($request->ids as $id) {
+      $team = Team::find($id);
+      try {
+        $team->delete();
+      } catch (\Exception $e) {
+        return ["status" => "danger", "message" => "Une erreur est survenue, réessayez plus tard."];
+      }
+    }
+
+    return response()->json([
+      'status' => 'alert',
+      'title' => 'Confirmation',
+      'content' => '<i class="fa fa-check-circle text-green"></i> La suppression a été effectuée avec succès',
+    ]);
   }
 
   public function get(Request $request) {
